Extract Identity into a separate domain
