+++ src/langium-model-server/identity/error.ts
