+++ src/langium-model-server/langium-model-server-module.ts
@@ -2,8 +2,8 @@ import type { LangiumSharedServices, Module } from 'langium'
 import { LmsDocumentHighlightProvider } from './lsp/document-highlight-provider'
 import { LmsLanguageServer } from './lsp/lms-language-server'
 import { LmsRenameProvider } from './lsp/rename-provider'
-import type { SemanticIdentity } from './semantic/identity'
-import type { IdentityIndex } from './semantic/identity-index'
+import type { SemanticIdentity } from './identity/model'
+import type { IdentityIndex } from './identity'
 import type { LangiumModelServerDefaultServices, LangiumModelServerDefaultSharedServices, LangiumModelServerServices } from './services'
 import { DefaultLangiumSourceModelServer } from './lms/langium-model-server'
 import { DefaultLmsSubscriptions } from './lms/subscriptions'
