+++ src/langium-model-server/lms/langium-model-server.ts
@@ -1,8 +1,8 @@
 import * as fs from 'fs-extra'
 import * as http2 from 'http2'
 import path from 'path'
-import type { SemanticIdentity } from '../semantic/identity'
-import type { IdentityIndex } from '../semantic/identity-index'
+import type { SemanticIdentity } from '../identity/model'
+import type { IdentityIndex } from '../identity'
 import type { LangiumModelServerServices } from '../services'
 import type { LmsDocument } from '../workspace/documents'
 import { LangiumModelServerRouter } from './router'
