+++ src/langium-model-server/lms/model/update/update.ts
@@ -1,4 +1,4 @@
-import type * as id from '../../../semantic/identity'
+import type * as id from '../../../identity/model'
 import type { KeysOfType, ModelAttribute, OptionalKeys, PrimitiveModelAttribute } from '../../../utils/types'
 import type { ArrayUpdate } from './array-update'
