+++ src/langium-model-server/lms/router.ts
@@ -1,8 +1,8 @@
 import * as http2 from 'http2'
 import { promisify } from 'util'
 import { isPromise } from 'util/types'
-import type { SemanticIdentity } from '../semantic/identity'
-import type { IdentityIndex } from '../semantic/identity-index'
+import type { SemanticIdentity } from '../identity/model'
+import type { IdentityIndex } from '../identity'
 import type { LangiumModelServerAddedServices, LmsServices } from '../services'
 import type { LmsDocument } from '../workspace/documents'
 import type { CreationParams, EditingResult } from './model'
