+++ src/language-server/task-list/semantic/task-list-identity-index.ts
@@ -1,6 +1,6 @@
-import type { AstNodeSemanticIdentity, DerivativeSemanticIdentity, Renameable } from '../../../langium-model-server/semantic/identity'
-import { ModelUri } from '../../../langium-model-server/semantic/identity'
-import type { IdentityIndex } from '../../../langium-model-server/semantic/identity-index'
+import type { AstNodeSemanticIdentity, DerivativeSemanticIdentity, Renameable } from '../../../langium-model-server/identity/model'
+import { ModelUri } from '../../../langium-model-server/identity/model'
+import type { IdentityIndex } from '../../../langium-model-server/identity'
 import { ValueBasedMap, equal } from '../../../langium-model-server/utils/collections'
 import type { Model, Task, TransitionDerivativeName } from './task-list-identity'
 import { Transition } from './task-list-identity'
